Phase 7.1d: Migrate HighlightDeleteWrapper to plain CSS
- Created plain CSS implementation with delete confirmation dialog styles
- Migrated textStyle to plain CSS
- Maintained backward compatibility with .legacy.tsx wrapper
- Added className prop for extensibility

Lint stuff

snaps
