+++ src/app/content/highlights/components/SummaryPopup/HighlightDeleteWrapper.css
@@ -1,0 +1,26 @@
+/**
+ * HighlightDeleteWrapper Styles - Plain CSS version
+ */
+
+.highlight-delete-wrapper {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.9);
+}
+
+.highlight-delete-wrapper span {
+  /* textStyle base */
+  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 25px;
+  letter-spacing: -0.2;
+  color: #fff;
+  margin-right: 8px;
+}
